+++ helpers/MobileAppHelper.php
@@ -2,6 +2,7 @@
 
 namespace humhub\modules\fcmPush\helpers;
 
+use humhub\helpers\DeviceDetectorHelper;
 use Yii;
 use yii\helpers\Json;
 use yii\helpers\Url;
@@ -16,9 +17,9 @@ class MobileAppHelper
     public const SESSION_VAR_REGISTER_NOTIFICATION = 'mobileAppRegisterNotification';
     public const SESSION_VAR_UNREGISTER_NOTIFICATION = 'mobileAppUnregisterNotification';
 
-    public static function registerHideOpenerScript()
+    public static function registerHideOpenerScript(): void
     {
-        if (!static::isAppRequest()) {
+        if (!DeviceDetectorHelper::isAppRequest()) {
             return;
         }
 
@@ -28,9 +29,9 @@ class MobileAppHelper
         self::sendFlutterMessage($message);
     }
 
-    public static function registerShowOpenerScript()
+    public static function registerShowOpenerScript(): void
     {
-        if (!static::isAppRequest()) {
+        if (!DeviceDetectorHelper::isAppRequest()) {
             return;
         }
 
@@ -40,9 +41,9 @@ class MobileAppHelper
         self::sendFlutterMessage($message);
     }
 
-    public static function registerNotificationScript()
+    public static function registerNotificationScript(): void
     {
-        if (!static::isAppRequest()) {
+        if (!DeviceDetectorHelper::isAppRequest()) {
             return;
         }
 
@@ -51,9 +52,9 @@ class MobileAppHelper
         self::sendFlutterMessage($message);
     }
 
-    public static function unregisterNotificationScript()
+    public static function unregisterNotificationScript(): void
     {
-        if (!static::isAppRequest()) {
+        if (!DeviceDetectorHelper::isAppRequest()) {
             return;
         }
 
@@ -62,52 +63,9 @@ class MobileAppHelper
         self::sendFlutterMessage($message);
     }
 
-    public static function isAppRequest()
-    {
-        return (
-            (Yii::$app->request->headers->get('x-requested-with', null, true) === 'com.humhub.app') ||
-            (Yii::$app->request->headers->has('x-humhub-app'))
-        );
-    }
-
-    /**
-     * Determines whether the app is a branded app with custom firebase configuration.
-     * @return bool
-     */
-    public static function isAppWithCustomFcm(): bool
-    {
-        return (
-            Yii::$app->request->headers->has('x-humhub-app-bundle-id') &&
-            !str_contains(
-                Yii::$app->request->headers->get('x-humhub-app-bundle-id', '', true),
-                'com.humhub.app',
-            )
-        );
-
-    }
-
-    private static function sendFlutterMessage($msg)
+    private static function sendFlutterMessage($msg): void
     {
         Yii::$app->view->registerJs('if (window.flutterChannel) { window.flutterChannel.postMessage(\'' . $msg . '\'); }');
-    }
-
-    public static function isIosApp(): bool
-    {
-        return
-            static::isAppRequest()
-            && Yii::$app->request->headers->get('x-humhub-app-is-ios');
-    }
-
-    /**
-     * True if the mobile app supports multi instance to display the Opener landing page without logout for switching instance
-     *
-     * @since HumHub mobile app v1.0.124
-     */
-    public static function isMultiInstanceApp(): bool
-    {
-        return
-            static::isAppRequest()
-            && Yii::$app->request->headers->get('x-humhub-app-is-multi-instance');
     }
 
     /**
@@ -118,7 +76,7 @@ class MobileAppHelper
     public static function openerState(): bool
     {
         return
-            static::isAppRequest()
+            DeviceDetectorHelper::isAppRequest()
             && Yii::$app->request->headers->get('x-humhub-app-opener-state');
     }
 }