+++ services/DriverService.php
@@ -2,10 +2,10 @@
 
 namespace humhub\modules\fcmPush\services;
 
+use humhub\helpers\DeviceDetectorHelper;
 use humhub\modules\fcmPush\driver\DriverInterface;
 use humhub\modules\fcmPush\driver\Fcm;
 use humhub\modules\fcmPush\driver\Proxy;
-use humhub\modules\fcmPush\helpers\MobileAppHelper;
 use humhub\modules\fcmPush\models\ConfigureForm;
 
 class DriverService
@@ -66,7 +66,7 @@ class DriverService
 
     public function getMobileAppDriver(): ?DriverInterface
     {
-        if (MobileAppHelper::isAppWithCustomFcm()) {
+        if (DeviceDetectorHelper::isAppWithCustomFcm()) {
             return $this->getConfiguredDriverByType(Fcm::class);
         }