+++ docs/CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 2.1.4 (Unreleased)
 -----------------------
 - Enh: Add "Open native console" button for the mobile app
+- Chg: The minimal HumHub version is now 1.17
+- Chg: Use the new `DeviceDetectorHelper` helper instead of `MobileAppHelper`
 
 2.1.3 (January 6, 2024)
 -----------------------